+++ src/seeds/seed-roles.ts
@@ -1,31 +1,36 @@
 // import { Role } from "@/entities";
 
-import type { User, Workspace } from "@/entities";
+import type { RoleRoute, User, Workspace } from "@/entities";
 import { Role } from "@/entities";
 import { DB } from "@/modules/api/DB";
 
 // seed default roles of a workspace
 export const seedDefaultRoles = async (workspace: Workspace, owner: User) => {
 	// ADMIN
-	const adminRoleDto = new Role();
-	adminRoleDto.name = "Administrator";
-	adminRoleDto.routes = [{ route: "*", permissions: ["full"] }];
-	adminRoleDto.workspace = workspace._id;
-	adminRoleDto.type = "admin";
+	let adminRole: Role = await DB.findOne<Role>("role", { type: "admin", workspace: workspace._id });
 
-	const adminRole = await DB.create<Role>("role", adminRoleDto);
-	if (adminRole) console.log(`Workspace "${workspace.name}" > Created default admin role :>> `, adminRoleDto.name);
+	if (!adminRole) {
+		const adminRoleDto = new Role();
+		adminRoleDto.name = "Administrator";
+		adminRoleDto.routes = [{ route: "*", permissions: ["full"] }];
+		adminRoleDto.workspace = workspace._id;
+		adminRoleDto.type = "admin";
+
+		adminRole = await DB.create<Role>("role", adminRoleDto);
+		if (adminRole) console.log(`Workspace "${workspace.name}" > Created default admin role :>> `, adminRoleDto.name);
+	}
 
 	// assign admin role to the "owner" user
 	const userRoles = owner.roles || [];
-	userRoles.push(adminRole);
-	const [user] = await DB.update<User>("user", { _id: owner._id }, { roles: userRoles });
-	console.log(`Workspace "${workspace.name}" > User "${user.name}" is now an administrator.`);
+	if (!userRoles.map((r) => r.toString()).includes(adminRole._id.toString())) {
+		userRoles.push(adminRole);
+		const [user] = await DB.update<User>("user", { _id: owner._id }, { roles: userRoles });
+		console.log(`Workspace "${workspace.name}" > User "${user.name}" is now an administrator.`);
+	}
 
 	// MEMBER
-	const memberRoleDto = new Role();
-	memberRoleDto.name = "Member";
-	memberRoleDto.routes = [
+	let memberRole: Role = await DB.findOne<Role>("role", { type: "member", workspace: workspace._id });
+	const memberRoleRoutes: RoleRoute[] = [
 		{ route: "*", permissions: ["own", "read"] },
 		{ route: "/api/v1/deploy", permissions: ["read", "create", "update"] },
 		{ route: "/api/v1/domain", permissions: ["read", "create", "update"] },
@@ -35,29 +40,60 @@ export const seedDefaultRoles = async (workspace: Workspace, owner: User) => {
 		{ route: "/api/v1/app/environment/variables", permissions: ["full"] },
 		{ route: "/api/v1/build/start", permissions: ["full"] },
 		{ route: "/api/v1/build/stop", permissions: ["full"] },
-		{ route: "/api/v1/user/join-workspace", permissions: ["update"] },
 		{ route: "/api/v1/release", permissions: ["own", "read", "create", "update"] },
 		{ route: "/api/v1/release/from-build", permissions: ["own", "read", "create", "update"] },
 		{ route: "/api/v1/release/preview", permissions: ["own", "read", "create", "update"] },
+		{ route: "/api/v1/git/public-key", permissions: [] },
+		{ route: "/api/v1/git/ssh/create", permissions: [] },
+		{ route: "/api/v1/git/ssh/generate", permissions: [] },
+		{ route: "/api/v1/git/ssh/verify", permissions: [] },
+		{ route: "/api/v1/user/join-workspace", permissions: ["update"] },
 		{ route: "/api/v1/role", permissions: ["read"] },
-		{ route: "/api/v1/api_key", permissions: ["read"] },
+		{ route: "/api/v1/api_key", permissions: [] },
 		{ route: "/api/v1/service_account", permissions: ["read"] },
 	];
-	memberRoleDto.workspace = workspace._id;
-	memberRoleDto.type = "member";
 
-	const memberRole = await DB.create<Role>("role", memberRoleDto);
-	if (memberRole) console.log(`Workspace "${workspace.name}" > Created default member role :>> `, memberRoleDto.name);
+	if (!memberRole) {
+		const memberRoleDto = new Role();
+		memberRoleDto.name = "Member";
+		memberRoleDto.routes = memberRoleRoutes;
+		memberRoleDto.workspace = workspace._id;
+		memberRoleDto.type = "member";
+
+		memberRole = await DB.create<Role>("role", memberRoleDto);
+		if (memberRole) console.log(`Workspace "${workspace.name}" > Created default member role :>> `, memberRoleDto.name);
+	} else {
+		// compare routes & permissions, if it doesn't match -> update!
+		const defaultMemberRoleRoutes = memberRoleRoutes.map((r) => `${r.route}:${r.permissions.join(",")}`).join("|");
+		const dbMemberRoleRoutes = memberRole.routes.map((r) => `${r.route}:${r.permissions.join(",")}`).join("|");
+		if (defaultMemberRoleRoutes !== dbMemberRoleRoutes) {
+			[memberRole] = await DB.update<Role>("role", { _id: memberRole._id }, { routes: memberRoleRoutes });
+			console.log(`Workspace "${workspace.name}" > Updated default member role!`);
+		}
+	}
 
 	// MODERATOR
-	const moderatorRoleDto = new Role();
-	moderatorRoleDto.name = "Moderator";
-	moderatorRoleDto.routes = [{ route: "*", permissions: ["own", "read", "create", "update"] }];
-	moderatorRoleDto.workspace = workspace._id;
-	moderatorRoleDto.type = "moderator";
+	let moderatorRole: Role = await DB.findOne<Role>("role", { type: "moderator", workspace: workspace._id });
+	const moderatorRoleRoutes: RoleRoute[] = [{ route: "*", permissions: ["own", "read", "create", "update"] }];
 
-	const moderatorRole = await DB.create<Role>("role", moderatorRoleDto);
-	if (moderatorRole) console.log(`Workspace "${workspace.name}" > Created default moderator role :>> `, moderatorRole.name);
+	if (!moderatorRole) {
+		const moderatorRoleDto = new Role();
+		moderatorRoleDto.name = "Moderator";
+		moderatorRoleDto.routes = moderatorRoleRoutes;
+		moderatorRoleDto.workspace = workspace._id;
+		moderatorRoleDto.type = "moderator";
+
+		moderatorRole = await DB.create<Role>("role", moderatorRoleDto);
+		if (moderatorRole) console.log(`Workspace "${workspace.name}" > Created default moderator role :>> `, moderatorRole.name);
+	} else {
+		// compare routes & permissions, if it doesn't match -> update!
+		const defaultModRoleRoutes = moderatorRoleRoutes.map((r) => `${r.route}:${r.permissions.join(",")}`).join("|");
+		const dbModRoleRoutes = moderatorRole.routes.map((r) => `${r.route}:${r.permissions.join(",")}`).join("|");
+		if (defaultModRoleRoutes !== dbModRoleRoutes) {
+			[moderatorRole] = await DB.update<Role>("role", { _id: moderatorRole._id }, { routes: moderatorRoleRoutes });
+			console.log(`Workspace "${workspace.name}" > Updated default moderator role!`);
+		}
+	}
 
 	return [adminRole, memberRole, moderatorRole];
 };